requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pytest
python-dotenv
sentence-transformers
crewai
openai
